// src/app/model/task.model.ts
export class Task {
    $key: string;
    tool: string; 
    in_use_by: string; 
    task: string;
    lat: number;
    lng: number;

   /* constructor(
        public $key: string,
        public  tool: string, 
        public  in_use_by: string, 
        public  task: string,
        public  lat: number,
        public  lng: number){
      
    }*/
}